+++ docs/source/change_log.rst
@@ -20,6 +20,7 @@ Version 1.9.x
    * Upgrade AdvancedHTTPServer to v2.0.11 to support async SSL handshakes
    * Support using an include directive in the server configuration file
    * Added a ``request-handle`` signal for custom HTTP request handlers
+   * Removed ``address`` support from the server config in favor of ``addresses``
 
 Version 1.8.0
 ^^^^^^^^^^^^^